Added tests for Training Address
